refactor(cvtservice): extract deployment_safety.go from god-file
Extract CanIDeploy + filterChangesForConsumer into a dedicated file.
Pure file movement, no behavior change.

Part 3 of 4 in the validator_service.go split (issue #107).
